requirements.txt
#
# Copyright 2018-2020 VMware, Inc.
# SPDX-License-Identifier: BSD-2-Clause
#

# Automatically generated by https://github.com/damnever/pigar.

# powerbi-repo/Objects/Config.py: 3
# powerbi-repo/Scripts/adam_test_template.py: 5
# powerbi-repo/adam_edge_wifi_usage.py: 7
# powerbi-repo/gateway_script.py: 8
# powerbi-repo/inventory_sla.py: 8
# powerbi-repo/powerbi_main_script.py: 15
PyYAML == 5.3.1

# powerbi-repo/gateway_script.py: 15
# powerbi-repo/powerbi_main_fun.py: 21
certifi == 2020.6.20

# powerbi-repo/gateway_script.py: 17
# powerbi-repo/powerbi_main_fun.py: 22
geoip2 == 4.0.1

# powerbi-repo/gateway_script.py: 16
# powerbi-repo/powerbi_main_fun.py: 25
geopy == 2.0.0

# powerbi-repo/Functions/sql_upserts.py: 4
# powerbi-repo/Scripts/adam_test_template.py: 4
# powerbi-repo/adam_edge_wifi_usage.py: 6
# powerbi-repo/fun_mysql_inserts.py: 6
# powerbi-repo/fun_mysql_query.py: 5
# powerbi-repo/gateway_script.py: 1
# powerbi-repo/inventory_sla.py: 6
# powerbi-repo/powerbi_main_fun.py: 26
# powerbi-repo/powerbi_main_script.py: 9
mysql_connector_python == 8.0.21

# powerbi-repo/Functions/helpers.py: 6
python_dateutil == 2.8.1

# powerbi-repo/Functions/vco_calls.py: 7,8
# powerbi-repo/UnitTests/test_vco_calls.py: 3,4
# powerbi-repo/VCOClient.py: 5
# powerbi-repo/gateway_script.py: 11,19
# powerbi-repo/inventory_sla.py: 9
# powerbi-repo/powerbi_main_fun.py: 24,27
# powerbi-repo/temp/VCOClient.py: 31
requests == 2.24.0

# powerbi-repo/powerbi_main_fun.py: 28
slack_webhook == 1.0.3

# powerbi-repo/gateway_script.py: 14
# powerbi-repo/inventory_sla.py: 7
urllib3 == 1.25.10
